Add files via upload
This code is a Python script that uses the snscrape library to scrape tweets from Twitter containing a specified keyword within a specified date range. The user inputs the keyword, start and end dates, and number of tweets to scrape through a Streamlit interface. The scraped tweets are then displayed in a pandas dataframe and the user can choose to upload the data to MongoDB and download the data in either CSV or JSON format. The code is suitable for anyone who wants to extract tweets related to a specific topic for analysis or research purposes. It can be added to a GitHub project as a standalone script or as part of a larger data scraping or analysis project.